A few paks i got working
QUAKE has a warning message for save states not working included. Shaun has added a new method for warning screens in the next release, so maybe the launch script would have to be modified by then.
Super Game Boy uses the mGBA core, for which it is definitely powerful enough.
Supafaust is an alternative to the snes9x core, but in my opinion Supafaust runs way better, especially on SH-1 and SFX games.
CPS1 works without a hitch.
Cave Story is supposedly going to be added as a native game soon, so i didn't include it here.