Fix Pico-8? (For now)
Thanks to SunClutch in the RGH Discord for pointing the issue out to me. LibSDL no longer required for binary to work with sound and proper rotation. Downside is it is not the latest release. Might need to recompile against shaun's SDL if I want it to work on MiniUI...